test/integration.sh: use client_secret_basic auth
When no OAuth2 client auth style is specified, the Golang RetrieveToken
method will try to get a token using the client_secret_basic method
first and then, if it fails, using the client_secret_post method. This
resulted in a transient error when opa-ams would fetch its token from
Hydra. This commit changes the auth style in hydra for the opa-ams
client to cleint_secret_basic so that the first request succeeds and no
transient error is logged.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>